# contracts/openapi.yaml
...
openapi: "3.1.0"
info:
  title: "材料实验助手 API"
  version: "1.0.0"
  description: |-
    当前后端实现的接口清单；维护依据为 backend 中的 Controller、Service、响应模型和安全配置。
    普通成功响应使用 data 和 request_id；分页另有 meta。空 data 按当前 NON_NULL 配置省略。
    204 和文件正文不使用成功信封。错误返回 application/json，包含业务 Problem Details 兼容对象、精简安全错误或内部错误；不能假定所有错误都有 detail、instance 或 request_id。

    前端使用同源 Session（默认 Cookie 为 JSESSIONID）。先获取 /auth/csrf 并保留 csrftoken，写请求携带匹配的 X-CSRFToken；登录也受 CSRF 校验。当前 SecurityConfig 另启用了 HTTP Basic。
    项目和实验读取按当前组织限定数据范围，普通用户另受项目、负责人或实验参与关系约束。创建实验或变更实验关联项目时先校验项目属于当前组织；超级管理员仅跳过项目管理成员检查。平台管理员仅能开通与查看组织目录，不获得其他组织的业务数据读取权。项目和实验受版本保护的操作使用 ETag / If-Match；当前没有 Idempotency-Key 去重实现。

    时间字段沿用现有 API 的 ISO 8601 表示，带时间的响应为含偏移量的日期时间。规范要求客户端展示及人工记录统一到 Asia/Shanghai，但当前页面部分格式化使用浏览器默认时区，尚未统一。项目和实验计划时间省略偏移量时服务端按 Asia/Shanghai 解释；调用方仍应优先提交含偏移量的日期时间。

    能力边界：ready 检查应用启动就绪及数据库连接；项目文档和实验附件正文保存在 PostgreSQL，文档 preview 返回原件。RustFS 存储和 LibreOffice 服务端转换尚未接入。
    实验参与人可随创建或 PATCH 写入；项目成员仅可读取，创建时加入创建者。开发初始化不能替代生产组织、角色和管理员引导；本契约不表示已完成生产部署验收。
servers:
  -
    url: "/api/v1"
    description: "同源 API"
...
security:
  -
    cookieAuth: []
  -
    basicAuth: []
paths:
...
  "/dashboard":
    get:
      operationId: "getDashboard"
      tags: ["Dashboard"]
      summary: "获取当前可见项目和实验统计"
      description: |-
        要求 project.read 权限。趋势按 Asia/Shanghai 当前日期生成最近 30 个日期。
        当前 project_id 仅回显到 trend.selected_project_id，没有参与统计查询，不应依赖它筛选趋势。
      parameters:
        -
          name: "project_id"
          in: "query"
          schema:
            type: "string"
            format: "uuid"
          description: "当前仅回显的项目选择值"
      responses:
        "200":
          description: "工作台统计"
          content:
            "application/json":
              schema:
                allOf:
                  -
                    "$ref": "#/components/schemas/Envelope"
                  -
                    type: "object"
                    required: ["data"]
                    properties:
                      data:
                        "$ref": "#/components/schemas/Dashboard"
        default:
          "$ref": "#/components/responses/ApiError"
  "/projects":
    get:
      operationId: "listProjects"
      tags: ["Projects"]
      summary: "分页查询当前用户可见项目"
      description: "要求 project.read 权限；普通用户按项目负责人或成员身份限定数据范围。"
      parameters:
        -
          "$ref": "#/components/parameters/Page"
        -
          "$ref": "#/components/parameters/PageSize"
        -
          "$ref": "#/components/parameters/Search"
        -
          name: "status"
          in: "query"
          schema:
            type: "string"
          description: "状态筛选；running 兼容映射为 active，ended 聚合 completed 和 archived，其余非空值按原值筛选"
      responses:
        "200":
          description: "项目分页列表"
          content:
            "application/json":
              schema:
                allOf:
                  -
                    "$ref": "#/components/schemas/PageEnvelope"
                  -
                    type: "object"
                    required: ["data"]
                    properties:
                      data:
                        type: "array"
                        items:
                          "$ref": "#/components/schemas/Project"
        default:
          "$ref": "#/components/responses/ApiError"
    post:
      operationId: "createProject"
      tags: ["Projects"]
      summary: "创建项目并加入创建者为管理成员"
      description: |-
        要求 project.create 和返回资源所需的 project.read 权限；负责人须为当前组织有效用户。
        当前不处理 members 或 member_ids，不实现 Idempotency-Key 去重；重试可能创建多个项目。
      security:
        -
          cookieAuth: []
          csrfCookie: []
          csrfHeader: []
        -
          basicAuth: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              allOf:
                -
                  "$ref": "#/components/schemas/ProjectWrite"
                -
                  type: "object"
                  required: ["name"]
      responses:
        "201":
          "$ref": "#/components/responses/Project"
        default:
          "$ref": "#/components/responses/ApiError"
  "/projects/{projectKey}":
    parameters:
      -
        "$ref": "#/components/parameters/ProjectKey"
    get:
      operationId: "getProject"
      tags: ["Projects"]
      summary: "获取可见项目详情及版本"
      description: "要求 project.read 权限；项目不存在或不在可见范围时返回 404。"
      responses:
        "200":
          "$ref": "#/components/responses/Project"
        default:
          "$ref": "#/components/responses/ApiError"
    patch:
      operationId: "updateProject"
      tags: ["Projects"]
      summary: "按资源版本更新项目"
      description: |-
        要求 project.update、project.read 和项目管理范围。未提交字段保留原值；当前不处理成员写入。
        已归档项目不能继续更新，版本不匹配或已归档时返回 412。
        objectives 和 milestones 整体替换，显式 null 写为空数组。
        计划时间省略或显式 null 均保留原值，当前没有清空已有计划时间的语义。
      security:
        -
          cookieAuth: []
          csrfCookie: []
          csrfHeader: []
        -
          basicAuth: []
          csrfCookie: []
          csrfHeader: []
      parameters:
        -
          "$ref": "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          "application/json":
            schema:
              "$ref": "#/components/schemas/ProjectWrite"
      responses:
        "200":
          "$ref": "#/components/responses/Project"
        "412":
          "$ref": "#/components/responses/ApiError"
        "428":
          "$ref": "#/components/responses/ApiError"
        default:
          "$ref": "#/components/responses/ApiError"
  "/projects/{projectKey}/archive":
    parameters:
      -
        "$ref": "#/components/parameters/ProjectKey"
    post:
      operationId: "archiveProject"
      tags: ["Projects"]
      summary: "归档项目"
      description: "要求 project.archive、project.read 和项目管理范围；已归档或版本不匹配返回 412。"
      security:
        -
          cookieAuth: []
          csrfCookie: []
          csrfHeader: []
        -
          basicAuth: []
          csrfCookie: []
          csrfHeader: []
      parameters:
        -
          "$ref": "#/components/parameters/IfMatch"
      responses:
        "200":
          "$ref": "#/components/responses/Project"
        "412":
          "$ref": "#/components/responses/ApiError"
        "428":
          "$ref": "#/components/responses/ApiError"
        default:
          "$ref": "#/components/responses/ApiError"
  "/projects/{projectKey}/follow":
    parameters:
      -
        "$ref": "#/components/parameters/ProjectKey"
    post:
      operationId: "followProject"
      tags: ["Projects"]
      summary: "关注可见项目"
      description: "要求 project.read 权限及项目可见性。"
      security:
        -
          cookieAuth: []
          csrfCookie: []
          csrfHeader: []
        -
          basicAuth: []
          csrfCookie: []
          csrfHeader: []
      responses:
        "200":
          "$ref": "#/components/responses/Follow"
        default:
          "$ref": "#/components/responses/ApiError"
    delete:
      operationId: "unfollowProject"
      tags: ["Projects"]
      summary: "取消关注可见项目"
      description: "要求 project.read 权限及项目可见性。"
      security:
        -
          cookieAuth: []
          csrfCookie: []
          csrfHeader: []
        -
          basicAuth: []
          csrfCookie: []
          csrfHeader: []
      responses:
        "200":
          "$ref": "#/components/responses/Follow"
        default:
          "$ref": "#/components/responses/ApiError"
  "/projects/{projectKey}/operation-logs":
    parameters:
      -
        "$ref": "#/components/parameters/ProjectKey"
    get:
      operationId: "listProjectOperationLogs"
      tags: ["Projects"]
      summary: "读取项目已有操作日志"
      description: "要求 project.read 权限及项目可见性；返回数据库已有日志，不承诺所有写操作均产生日志。"
      responses:
        "200":
          description: "项目操作日志列表"
          content:
            "application/json":
              schema:
                allOf:
                  -
                    "$ref": "#/components/schemas/Envelope"
                  -
                    type: "object"
                    required: ["data"]
                    properties:
                      data:
                        type: "array"
                        items:
                          "$ref": "#/components/schemas/ProjectOperationLog"
        default:
          "$ref": "#/components/responses/ApiError"
  "/projects/{projectKey}/documents":
    parameters:
      -
        "$ref": "#/components/parameters/ProjectKey"
    get:
      operationId: "listProjectDocuments"
      tags: ["Documents"]
      summary: "查询项目文档和分类统计"
      description: "要求 document.view、project.read 及项目可见性；列表不分页，分类总数不受筛选影响。"
      parameters:
        -
          name: "category"
          in: "query"
          schema:
            "$ref": "#/components/schemas/DocumentCategory"
        -
          "$ref": "#/components/parameters/Search"
        -
          name: "file_type"
          in: "query"
          schema:
            type: "string"
            enum: ["word","pdf","excel","powerpoint","image","text"]
          description: "按文件名扩展名筛选，不依据 MIME 类型"
        -
          name: "updated_range"
          in: "query"
          schema:
            type: "string"
            enum: ["week","month"]
          description: "week 为最近 7 天，month 为最近 30 天"
      responses:
        "200":
          description: "文档列表、项目总数、筛选总数和分类计数"
          content:
            "application/json":
              schema:
                "$ref": "#/components/schemas/DocumentListResponse"
        default:
          "$ref": "#/components/responses/ApiError"
    post:
      operationId: "uploadProjectDocument"
      tags: ["Documents"]
      summary: "上传文档元数据和原始正文"
      description: |-
        要求 document.upload、project.read 和项目管理范围。正文写入 PostgreSQL。
        服务层要求文件非空且不超过 20971520 字节；实际可上传大小还受服务器 multipart 与反向代理限制。
        当前没有内容格式鉴定、对象存储转存或服务端办公文档转换。
      security:
        -
          cookieAuth: []
          csrfCookie: []
          csrfHeader: []
        -
          basicAuth: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          "multipart/form-data":
            schema:
              type: "object"
              required: ["file","category","version_label"]
              properties:
                file:
                  type: "string"
                  format: "binary"
                  description: "非空原始文件"
                category:
                  "$ref": "#/components/schemas/DocumentCategory"
                version_label:
                  type: "string"
                  minLength: 1
                  description: "去除首尾空白后必须非空"
      responses:
        "201":
          description: "已保存的文档元数据"
          content:
            "application/json":
              schema:
                allOf:
                  -
                    "$ref": "#/components/schemas/Envelope"
                  -
                    type: "object"
                    required: ["data"]
                    properties:
                      data:
                        "$ref": "#/components/schemas/ProjectDocument"
        default:
          "$ref": "#/components/responses/ApiError"
  "/projects/{projectKey}/documents/{documentId}/content":
    parameters:
      -
        "$ref": "#/components/parameters/ProjectKey"
      -
        "$ref": "#/components/parameters/DocumentId"
    get:
      operationId: "getProjectDocumentContent"
      tags: ["Documents"]
      summary: "读取或下载项目文档原件"
      description: "要求 document.view、project.read 及项目可见性；文档或正文不存在返回 404。"
      parameters:
        -
          name: "download"
          in: "query"
          schema:
            type: "boolean"
            default: false
          description: "true 使用 attachment，false 使用 inline"
      responses:
        "200":
          "$ref": "#/components/responses/DocumentBinary"
        default:
          "$ref": "#/components/responses/ApiError"
  "/projects/{projectKey}/documents/{documentId}/preview":
    parameters:
      -
        "$ref": "#/components/parameters/ProjectKey"
      -
        "$ref": "#/components/parameters/DocumentId"
    get:
      operationId: "getProjectDocumentPreview"
      tags: ["Documents"]
      summary: "以内联方式读取文档原件"
      description: |-
        与 content 调用相同正文读取逻辑，Content-Disposition 固定为 inline。
        返回原始字节和保存的 MIME 类型（包括 image/webp 等），不会把办公文档转换为 PDF。
        浏览器或前端能否预览取决于实际文件格式及其预览组件。要求 document.view、project.read 及项目可见性。
      responses:
        "200":
          "$ref": "#/components/responses/DocumentBinary"
        default:
          "$ref": "#/components/responses/ApiError"
...
components:
  securitySchemes:
    cookieAuth:
      type: "apiKey"
      in: "cookie"
      name: "JSESSIONID"
      description: "Servlet Session 默认 Cookie 名；服务器配置可覆盖"
    basicAuth:
      type: "http"
      scheme: "basic"
      description: "当前 SecurityConfig 启用的 HTTP Basic；写请求仍受 CSRF 约束"
    csrfCookie:
      type: "apiKey"
      in: "cookie"
      name: "csrftoken"
      description: "先调用 GET /auth/csrf 获取并保留"
    csrfHeader:
      type: "apiKey"
      in: "header"
      name: "X-CSRFToken"
      description: "与当前 csrftoken Cookie 匹配的令牌"
  parameters:
    Page:
      name: "page"
      in: "query"
      schema:
        type: "integer"
        default: 1
      description: "页码从 1 开始，低于 1 按 1 处理"
    PageSize:
      name: "page_size"
      in: "query"
      schema:
        type: "integer"
        default: 20
      description: "每页数量，服务端归一化到 1 至 100"
    Search:
      name: "search"
      in: "query"
      schema:
        type: "string"
      description: "关键字模糊筛选，具体搜索列以接口查询实现为准"
...
    ProjectKey:
      name: "projectKey"
      in: "path"
      required: true
      schema:
        type: "string"
      description: "项目 UUID 字符串或 PRJ 业务编号"
    DocumentId:
      name: "documentId"
      in: "path"
      required: true
      schema:
        type: "string"
        format: "uuid"
      description: "项目文档主键"
...
    IfMatch:
      name: "If-Match"
      in: "header"
      required: true
      schema:
        type: "string"
        example: "\"1\""
      description: "使用最新 ETag 或带引号的 version 整数。整数解析失败返回 428，版本冲突返回 412；当前缺失请求头由通用异常处理兜底，不能承诺与格式错误同样返回 428。"
  responses:
    ApiError:
      description: "业务或参数错误为 Problem Details 兼容对象；安全过滤器返回 status/code/detail；未分类异常返回内部错误。通常 401 为未认证、403 为权限或 CSRF 失败、404 为资源不可见或不存在。当前媒体类型为 application/json，不能假定所有非法输入均为 400。"
      content:
        "application/json":
          schema:
            anyOf:
              -
                "$ref": "#/components/schemas/ProblemDetail"
              -
                "$ref": "#/components/schemas/SecurityError"
              -
                "$ref": "#/components/schemas/InternalError"
...
    Project:
      description: "项目详情及最新 ETag"
      headers:
        ETag:
          description: "带双引号的版本整数"
          schema:
            type: "string"
            example: "\"1\""
      content:
        "application/json":
          schema:
            allOf:
              -
                "$ref": "#/components/schemas/Envelope"
              -
                type: "object"
                required: ["data"]
                properties:
                  data:
                    "$ref": "#/components/schemas/Project"
    Follow:
      description: "当前用户的项目关注状态"
      content:
        "application/json":
          schema:
            allOf:
              -
                "$ref": "#/components/schemas/Envelope"
              -
                type: "object"
                required: ["data"]
                properties:
                  data:
                    type: "object"
                    required: ["is_followed"]
                    properties:
                      is_followed:
                        type: "boolean"
...
    DocumentBinary:
      description: "原始文件字节，无 JSON 信封。Content-Type 使用保存的上传 MIME 类型；无法解析时回退 application/octet-stream。包括 PDF、办公文档、image/webp 等实际上传类型，不保证转换或格式安全。"
      headers:
        Content-Disposition:
          description: "inline 或 attachment，并携带文件名"
          schema:
            type: "string"
        Content-Length:
          description: "原始正文的字节数"
          schema:
            type: "integer"
            format: "int64"
            minimum: 0
      content:
        "*/*":
          schema:
            type: "string"
            format: "binary"
  schemas:
    Envelope:
      type: "object"
      required: ["request_id"]
      properties:
        request_id:
          type: "string"
          description: "当前响应生成的标识"
          format: "uuid"
      description: "非空业务数据通过 data 返回；空 data 按当前 NON_NULL 配置省略。"
    PageEnvelope:
      allOf:
        -
          "$ref": "#/components/schemas/Envelope"
        -
          type: "object"
          required: ["meta"]
          properties:
            meta:
              "$ref": "#/components/schemas/PageMeta"
...
    ProjectWrite:
      type: "object"
      required: []
      properties:
        name:
          type: "string"
          minLength: 1
        project_type_code:
          type: "string"
          description: "创建时默认 research"
        description:
          type: "string"
          description: "创建时默认空串"
        current_stage:
          type: "string"
          description: "创建时默认方案设计"
        progress_percent:
          type: "integer"
          minimum: 0
          maximum: 100
          description: "创建时默认 0"
        status:
          type: "string"
          description: "创建时默认 not_started；当前服务层未限制枚举或校验项目状态迁移"
        owner_id:
          type: "string"
          description: "创建时缺省为当前用户"
          format: "uuid"
        objectives:
          type: ["array","null"]
          items:
            type: "string"
          description: "整体替换；创建时缺省为空数组；null 写为空数组"
        milestones:
          type: ["array","null"]
          items:
            "$ref": "#/components/schemas/ProjectMilestone"
          description: "整体替换；创建时缺省为空数组；null 写为空数组"
        planned_start_date:
          type: ["string","null"]
          description: "ISO 8601 日期时间；无偏移量时按 Asia/Shanghai 解释；省略或 null 保留现值"
          examples: ["2026-09-07T09:00:00+08:00","2026-09-07T09:00:00"]
        planned_end_date:
          type: ["string","null"]
          description: "与 planned_start_date 相同格式及更新语义"
      description: "当前服务读取的项目写字段；成员信息和响应统计不在写入范围。"
...
    Project:
      type: "object"
      required: ["id","organization_id","project_no","name","project_type_code","description","current_stage","progress_percent","document_count","experiment_count","data_resource_count","objectives","milestones","status","owner_id","owner_display_name","members","version","created_at","updated_at"]
      properties:
        id:
          type: "string"
          format: "uuid"
        organization_id:
          type: "string"
          format: "uuid"
        project_no:
          type: "string"
        name:
          type: "string"
        project_type_code:
          type: "string"
        description:
          type: "string"
        current_stage:
          type: "string"
        progress_percent:
          type: "integer"
        document_count:
          type: "integer"
        experiment_count:
          type: "integer"
          description: "项目表中的存储计数，不代表独立实时统计"
        data_resource_count:
          type: "integer"
          description: "保留统计字段，不表示存在数据资产 API"
        objectives:
          type: "array"
          items:
            type: "string"
        milestones:
          type: "array"
          items:
            "$ref": "#/components/schemas/ProjectMilestone"
        status:
          type: "string"
        owner_id:
          type: "string"
          format: "uuid"
        owner_display_name:
          type: "string"
        members:
          type: "array"
          items:
            "$ref": "#/components/schemas/ProjectMember"
        planned_start_date:
          type: "string"
          format: "date-time"
        planned_end_date:
          type: "string"
          format: "date-time"
        actual_end_at:
          type: "string"
          format: "date-time"
        archived_at:
          type: "string"
          format: "date-time"
        version:
          type: "integer"
          minimum: 1
        created_at:
          type: "string"
          format: "date-time"
        updated_at:
          type: "string"
          format: "date-time"
      description: "未赋值的可空日期时间字段按当前配置省略。"
    ProjectOperationLog:
      type: "object"
      required: ["id","action_type","description","changes","created_at"]
      properties:
        id:
          type: "string"
          format: "uuid"
        action_type:
          type: "string"
        description:
          type: "string"
        actor_display_name:
          type: "string"
        changes:
          type: "object"
          additionalProperties: true
        created_at:
          type: "string"
          format: "date-time"
    DocumentCategory:
      type: "string"
      enum: ["project_plan","literature","experiment_plan","stage_report","meeting_minutes","other"]
...
    DocumentListResponse:
      allOf:
        -
          "$ref": "#/components/schemas/Envelope"
        -
          type: "object"
          required: ["data","meta"]
          properties:
            data:
              type: "array"
              items:
                "$ref": "#/components/schemas/ProjectDocument"
            meta:
              type: "object"
              required: ["total","filtered_total","category_counts"]
              properties:
                total:
                  type: "integer"
                  format: "int64"
                  minimum: 0
                  description: "项目全部文档数"
                filtered_total:
                  type: "integer"
                  format: "int64"
                  minimum: 0
                  description: "满足当前筛选的文档数"
                category_counts:
                  type: "object"
                  required: ["project_plan","literature","experiment_plan","stage_report","meeting_minutes","other"]
                  additionalProperties:
                    type: "integer"
                    format: "int64"
                    minimum: 0
                  description: "项目全部文档按分类统计，零项也返回 0；不受本次筛选影响"
...
    Dashboard:
      type: "object"
      required: ["project_metrics","experiment_metrics","type_distribution","trend","active_projects"]
      properties:
        project_metrics:
          "$ref": "#/components/schemas/DashboardMetrics"
        experiment_metrics:
          "$ref": "#/components/schemas/DashboardMetrics"
        type_distribution:
          type: "array"
          items:
            type: "object"
            required: ["name","project_count","experiment_count"]
            properties:
              name:
                type: "string"
              project_count:
                type: "integer"
                format: "int64"
              experiment_count:
                type: "integer"
                format: "int64"
        trend:
          type: "object"
          required: ["dates","series","selected_project_id","project_options"]
          properties:
            dates:
              type: "array"
              items:
                type: "string"
                format: "date"
              minItems: 30
              maxItems: 30
            series:
              type: "array"
              items:
                type: "object"
                required: ["name","values"]
                properties:
                  name:
                    type: "string"
                  values:
                    type: "array"
                    items:
                      type: "integer"
                      format: "int64"
                    minItems: 30
                    maxItems: 30
            selected_project_id:
              type: "string"
              description: "project_id 查询参数回显；未提供时为空串，当前不改变统计数据"
            project_options:
              type: "array"
              items:
                type: "object"
                required: ["id","project_no","name"]
                properties:
                  id:
                    type: "string"
                    format: "uuid"
                  project_no:
                    type: "string"
                  name:
                    type: "string"
        active_projects:
          type: "array"
          items:
            "$ref": "#/components/schemas/DashboardProject"